\name{PenPC-package}
\alias{PenPC-package}
\alias{PenPC}
\docType{package}
\title{
Estimating directed Acyclic graphs from Gaussian Graphical Models
}
\description{
Estimating the skeletons of directed Acyclic graphs from Gaussian Graphical Models}
\author{
Min Jin Ha <minjin.ha@gmail.com>
}
\references{
Min Jin Ha, Wei Sun and Jichung Xie. PenPC: A two-step approach to estimate the skeletons of high dimensional directed acyclic graphs
}
\keyword{ package }
